refactor: move deepseek_v3 tokenizer patch from nemo_rl.__init__ to nemo_rl.models.policy.__init__
The patch is only needed by policy code paths that eventually call
AutoTokenizer.from_pretrained -- driver's get_tokenizer(), Megatron
worker's build_tokenizer, DTensor worker's tokenizer reconstruction.
All three share nemo_rl.models.policy as an import ancestor.

Placing it in nemo_rl.models.policy.__init__.py fires exactly once
per process when policy code is first imported. Non-policy consumers
(tests, utilities, environments) no longer mutate the global
transformers registries as a side effect of importing nemo_rl.

Bonus: setuptools build isolation reads nemo_rl.__version__ from
nemo_rl.__init__.py but never cascades into models/policy, so the
_is_build_isolation() guard is no longer needed at the call site.

Verified locally with 'from nemo_rl.models.policy import PolicyConfig':
deepseek_v3 removed from MODELS_WITH_INCORRECT_HUB_TOKENIZER_CLASS
and TOKENIZER_MAPPING_NAMES as expected.
